Fix: Norminette Issue

+++ philosophers/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: tsuchen <tsuchen@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/06 15:13:01 by tsuchen           #+#    #+#             */
-/*   Updated: 2024/08/08 19:11:48 by tsuchen          ###   ########.fr       */
+/*   Updated: 2024/08/08 19:44:22 by tsuchen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,11 +32,11 @@ typedef enum e_task
 
 typedef struct s_setup
 {
-	unsigned int	phils;
-	unsigned long	time_to_die;
-	unsigned long	time_to_eat;
-	unsigned long	time_to_sleep;
-	unsigned int	must_eat_times;
+	uint32_t		phils;
+	uint32_t		time_to_die;
+	uint32_t		time_to_eat;
+	uint32_t		time_to_sleep;
+	uint16_t		must_eat_times;
 	struct timeval	start;
 	struct timeval	*last_meal;
 	pthread_mutex_t	*mtx_fork;
@@ -52,18 +52,18 @@ typedef struct s_philo
 }	t_philo;
 
 /* utils libft functions */
-long	ft_atol(const char *nptr);
+long		ft_atol(const char *nptr);
 /* utils functions */
-unsigned long	get_time(struct timeval *ref);
-void	init_phil(t_philo *phil, int i, t_setup *setting);
-int		input_check(int ac, char *av[], t_setup *setting);
-void    init_mutex(t_setup *setting, pthread_mutex_t *mutexFork);
-void	destroy_mutex(t_setup *setting, pthread_mutex_t *mutexFork);
+uint32_t	get_time(struct timeval *ref);
+void		init_phil(t_philo *phil, int i, t_setup *setting);
+int			input_check(int ac, char *av[], t_setup *setting);
+void		init_mutex(t_setup *setting, pthread_mutex_t *mutexFork);
+void		destroy_mutex(t_setup *setting, pthread_mutex_t *mutexFork);
 /* error functions */
-int		ac_check(int ac);
+int			ac_check(int ac);
 /* activity functions */
-int		eating_with_fork(t_philo *philo);
-void	sleeping(t_philo *philo);
-void	thinking(t_philo *philo);
+int			eating_with_fork(t_philo *philo);
+void		sleeping(t_philo *philo);
+void		thinking(t_philo *philo);
 
 #endif
